+++ client/src/components/ModeSwitcher.svelte
@@ -48,7 +48,7 @@
       {#if $modesAllowed.length < 2}
         <span class="mode-switcher-label">{LABELS[$effectiveMode]}</span>
       {/if}
-      {#if $modesAllowed.length > 1}<svg class="icon menu-chevron"><use href="#icon-chevron-down"></use></svg>{/if}
+      {#if $modesAllowed.length > 1}<svg class="icon mode-switcher-caret"><use href="#icon-chevron-down"></use></svg>{/if}
     </button>
     {#if open && $modesAllowed.length > 1}
       <div class="dropdown-menu mode-switcher-menu" role="menu">